Run a complete parametric survival analysis

This function is a wrapper for the runPSM function intended to be run with the boot package. It enables running a complete parametric survival analysis for use when performing bootstrapping to explore uncertainty. By re-using random seeds for each bootstrap sample it is possible to maintain correlations across multiple endpoints.

bootPSM(data, i, ...)

Arguments

data
A data frame containing individual patient data for the relevant time to event outcomes. This is passed to the data argument of the runPSM function

i
Index used to select a sample within boot.

...
Additional parameters as used by runPSM

Value

The 'parameters_vector' object from the runPSM function.

'parameters_vector' is a vector which contains the coefficients for all of the flexsurv models specified. The column names are in the format 'modeltype.distribution.parameter.TreatmentName', for example, comshp.weibull.shape.Int refers to the shape parameter of the common shape weibull distribution for the intervention treatment and 'indshp.gengamma.mu.ref' refers to the mu parameter of the independent shape generalised gamma distribution for the reference treatment. Columns with 'TE' at the end are the treatment effect coefficients (applicable to the scale and shape parameters for independent shape models, applicable to the scale parameter only for the common shape model and not applicable for the separate model).

Details

For more details and examples see the package vignettes:

vignette("Fitting_models_in_R", package = "flexsurvPlus")

vignette("Fitting_models_in_R_bootstrap", package = "flexsurvPlus")

vignette("Survival_analysis_theory", package = "flexsurvPlus")

This function is intended to be used in conjunction with the boot function to return the statistic to be bootstrapped. In this case by performing parametric survival modelling using flexsurv and returning the parameters of the survival distributions. This is used as the 'statistic' argument in the boot function.